split SqliteDAO helper methods into helpers
