+++ workers/loc.api/sync/dao/helpers/index.js
@@ -1,0 +1,23 @@
+'use strict'
+
+const {
+  mixUserIdToArrData,
+  convertDataType
+} = require('./utils')
+const {
+  serializeVal,
+  deserializeVal
+} = require('./serialization')
+const getWhereQuery = require('./get-where-query')
+const getOrderQuery = require('./get-order-query')
+const getUniqueIndexQuery = require('./get-unique-index-query')
+
+module.exports = {
+  mixUserIdToArrData,
+  convertDataType,
+  serializeVal,
+  deserializeVal,
+  getWhereQuery,
+  getOrderQuery,
+  getUniqueIndexQuery
+}
